--- src/components/EventComponent/index.js
@@ -1,1 +1,0 @@
-export { default as EventComponent } from './EventComponent';
