+++ src/components/EventComponent/index.ts
@@ -1,0 +1,1 @@
+export * from './EventComponent';
